[resources] Add StringArrayResource
