+++ components/resources/library/src/commonMain/kotlin/org/jetbrains/compose/resources/StringArrayResources.kt
@@ -1,0 +1,66 @@
+package org.jetbrains.compose.resources
+
+import androidx.compose.runtime.*
+import kotlinx.coroutines.*
+import kotlinx.coroutines.sync.Mutex
+import kotlinx.coroutines.sync.withLock
+import org.jetbrains.compose.resources.plural.PluralCategory
+import org.jetbrains.compose.resources.plural.PluralRuleList
+import org.jetbrains.compose.resources.vector.xmldom.Element
+import org.jetbrains.compose.resources.vector.xmldom.NodeList
+
+/**
+ * Represents a string array resource in the application.
+ *
+ * @param id The unique identifier of the resource.
+ * @param key The key used to retrieve the string array resource.
+ * @param items The set of resource items associated with the string array resource.
+ */
+@OptIn(InternalResourceApi::class)
+@ExperimentalResourceApi
+@Immutable
+class StringArrayResource
+@InternalResourceApi constructor(id: String, val key: String, items: Set<ResourceItem>) : Resource(id, items)
+
+/**
+ * Retrieves a list of strings using the specified string array resource.
+ *
+ * @param resource The string array resource to be used.
+ * @return A list of strings representing the items in the string array.
+ *
+ * @throws IllegalStateException if the string array with the given ID is not found.
+ */
+@ExperimentalResourceApi
+@Composable
+fun stringArrayResource(resource: StringArrayResource): List<String> {
+    val resourceReader = LocalResourceReader.current
+    val array by rememberResourceState(resource, { emptyList() }) { env ->
+        loadStringArray(resource, resourceReader, env)
+    }
+    return array
+}
+
+/**
+ * Loads a list of strings using the specified string array resource.
+ *
+ * @param resource The string array resource to be used.
+ * @return A list of strings representing the items in the string array.
+ *
+ * @throws IllegalStateException if the string array with the given ID is not found.
+ */
+@ExperimentalResourceApi
+suspend fun getStringArray(resource: StringArrayResource): List<String> =
+    loadStringArray(resource, DefaultResourceReader, getResourceEnvironment())
+
+@OptIn(ExperimentalResourceApi::class)
+private suspend fun loadStringArray(
+    resource: StringArrayResource,
+    resourceReader: ResourceReader,
+    environment: ResourceEnvironment
+): List<String> {
+    val path = resource.getPathByEnvironment(environment)
+    val keyToValue = getParsedStrings(path, resourceReader)
+    val item = keyToValue[resource.key] as? StringItem.Array
+        ?: error("String array ID=`${resource.key}` is not found!")
+    return item.items
+}
